requirements.txt
python-dotenv
ipykernel
agno 
openai
pypdf
lancedb 
tantivy
pandas
yfinance
streamlit
duckduckgo-search